requirements.txt
streamlit>=1.30.0,<2.0.0
PyMuPDF>=1.23.0,<2.0.0
pytesseract>=0.3.10,<1.0.0
Pillow>=10.0.0
sentence-transformers>=2.2.0,<3.0.0
chromadb>=0.4.0,<1.0.0
httpx>=0.25.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
psutil>=5.9.0,<6.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-mock>=3.12.0,<4.0.0
